git.py: Use constants in target_protection.py to overhaul branch checking, including implementing it for remotes with the SUBTREE layout (SOFTWARE-6078)
When building into a Koji target, the name of the target is checked
against the `koji_target_re` patterns in RESTRICTED_TARGETS.  If the
target matches one of them, then the remote being built from is checked
against the remotes in listed in that object.

If the remote has the LEGACY layout, the Git branch is checked against
the git_branch_re pattern; if it does not match, the check fails.

If the remote has the SUBTREE layout, the parent directory of the
package is checked against the `subtree_branch_re` pattern; if it does
not match, the check fails.  TODO: the Git branch is not checked for
remotes with the SUBTREE layout, though it probably should be.  Despite
this, builds from Git are rejected if the RestrictedTarget does not have
a `git_branch_re` pattern, which is definitely a bug.

When a check fails, VCSError is raised.

To simplify the code, the invocations of `git` are now run against the
package directory instead of the "top" directory -- finding the top dir
in the first place required yet another invocation of git, so it was
just annoying.  Now we need to use the `-C` flag which requires Git 2.0+
but an earlier commit, ac94e88ff0425c50c47cd94916b833fafa9355e6, adds
a check for that.