svn.py: Use constants in target_protection.py to overhaul branch checking (SOFTWARE-6078)
I was able to eliminate a lot of the code by requiring that the SVN
branch and the Koji target match, as opposed to forbidding certain
branches from building into certain targets.  For example, you used to
be able to build into any target you wanted from `branches/matyas`.
Now, building into one of the `23-main` targets requires the build to be
done from `branches/23-main`.

*This is a behavior change* but IMO the semantics are better this way,
not just from a code simplicity perspective.  As previously, this
restriction (and target protection in general) only applies to
non-scratch builds.